Update use signUpObject
