+++ shared/common/src/components/SignUp/Pagination/Page1/index.tsx
@@ -63,7 +63,6 @@ const Page1 = ({
           ? !value.length &&
             isScrollContainer && (
               <SignUpScrollContainer
-                value={value}
                 idx={idx}
                 placeholder={placeholder}
                 obj={page1Obj}
@@ -75,7 +74,6 @@ const Page1 = ({
             !value.length &&
             isScrollContainer && (
               <SignUpScrollContainer
-                value={value}
                 idx={idx}
                 placeholder={placeholder}
                 obj={page1Obj}
