+++ shared/common/src/constants/signUpObject.ts
@@ -1,0 +1,95 @@
+export const belongs = ['학교', '외부']
+export const insideJob = ['학생', '취업동아리 선생님']
+export const outsideJob = ['기업 강사', '대학 교수', '뽀짝 선생님', '유관기관']
+
+export const school: string[] = [
+  '금파공업고',
+  '전남공업고',
+  '광주공업고',
+  '전남여자상업고',
+  '광주자연과학고',
+  '광주전자공업고',
+  '서진여자고',
+  '숭의과학기술고',
+  '동일미래과학고',
+  '광주여자상업고',
+  '송원여자상업고',
+  '광주자동화설비마이스터고',
+  '광주소프트웨어마이스터고',
+]
+
+export const club: Record<string, string[]> = {
+  금파공업고: [
+    '레프리',
+    '블라썸(blossom)',
+    '유선통신',
+    '천기꿈나무',
+    '어썸(awesome)',
+    '다이나믹(Dynamic)',
+    '금호로80 베이커리',
+  ],
+  전남공업고: [
+    '진짜기계',
+    '핫앤쿨',
+    '에너지지키미',
+    '라온하제',
+    '스카이드론',
+    '그린라이트',
+  ],
+  광주공업고: [
+    'SMART JOB PROJECT',
+    '나의 미래는 내가 주인공이다!',
+    '설비의 달인',
+    '특수용접 화이팅',
+    '전기가 미래다',
+    '전자 어벤져스',
+    '전자 히어로스',
+    'Civil 마스터',
+    '건축연구소',
+  ],
+  전남여자상업고: [],
+  광주자연과학고: [
+    'DCT',
+    '뉴쿡',
+    '베이커리 카페 CEO',
+    '우(리들의) 아(름다운) 행(복)',
+  ],
+  광주전자공업고: [
+    '감성기계',
+    '열정 그 자체',
+    'ACT',
+    'ECT',
+    'Tesla',
+    '발자국',
+    '"M lab"= 메이커(maker) 연구소',
+  ],
+  서진여자고: [],
+  숭의과학기술고: [
+    '서전트스나이퍼',
+    '카-페인팅',
+    '드림온',
+    '볼트와 암페어',
+    '크로스핏마스터',
+    '비상',
+    '캐치어드론',
+    '내빵네빵',
+    '카페바리',
+    '쿠킹마스터즈',
+  ],
+  동일미래과학고: [
+    '놀고잡고(놀GO잡GO)',
+    '믿고잡고',
+    '따고잡고',
+    '쓰고잡고',
+    '하고잡고',
+  ],
+  광주여자상업고: ['금융실무', '소개팅', '취사모'],
+  송원여자상업고: ['건강지킴이', '미용서비스', '뷰티아트', '클로즈업'],
+  광주자동화설비마이스터고: [
+    'HMI동아리',
+    '마취제',
+    '빛go job go',
+    '취업진로 동아리',
+  ],
+  광주소프트웨어마이스터고: ['dev GSM'],
+}
